+++ caldir-provider-google/src/commands/list_events.rs
@@ -6,10 +6,10 @@ use google_calendar::types::OrderBy;
 use serde::Deserialize;
 
 use crate::DEFAULT_CALENDAR_ID;
-use crate::commands::authenticate::redirect_uri;
 use crate::config;
-use crate::google::actions::get_valid_tokens;
-use crate::google::from_google::from_google_event;
+use crate::google_auth::get_valid_tokens;
+use crate::google_auth::redirect_uri;
+use crate::transformer::from_google_event;
 
 #[derive(Debug, Deserialize)]
 struct ListEventsParams {
