+++ caldir-provider-google/src/commands/update_event.rs
@@ -2,10 +2,9 @@ use anyhow::Result;
 use google_calendar::{Client, types::SendUpdates};
 use serde::Deserialize;
 
-use crate::commands::authenticate::redirect_uri;
 use crate::config;
-use crate::google::actions::get_valid_tokens;
-use crate::google::{from_google::from_google_event, to_google::to_google_event};
+use crate::google_auth::{get_valid_tokens, redirect_uri};
+use crate::transformer::{from_google_event, to_google_event};
 use crate::{DEFAULT_CALENDAR_ID, types};
 
 #[derive(Debug, Deserialize)]
