--- caldir-provider-google/src/google/mod.rs
@@ -1,3 +1,0 @@
-pub mod actions;
-pub mod from_google;
-pub mod to_google;
