+++ caldir-provider-google/src/transformer/to_google.rs
@@ -1,31 +1,4 @@
-use crate::types::{Event, EventStatus, EventTime, ParticipationStatus, Transparency};
-use google_calendar::types::{EventAttendee, EventDateTime, EventReminder, Reminders};
-
-/// Convert EventTime to Google's EventDateTime
-fn event_time_to_google(time: &EventTime) -> EventDateTime {
-    match time {
-        EventTime::Date(d) => EventDateTime {
-            date: Some(*d),
-            date_time: None,
-            time_zone: String::new(),
-        },
-        EventTime::DateTimeUtc(dt) => EventDateTime {
-            date: None,
-            date_time: Some(*dt),
-            time_zone: String::new(),
-        },
-        EventTime::DateTimeFloating(dt) => EventDateTime {
-            date: None,
-            date_time: Some(dt.and_utc()),
-            time_zone: String::new(),
-        },
-        EventTime::DateTimeZoned { datetime, tzid } => EventDateTime {
-            date: None,
-            date_time: Some(datetime.and_utc()),
-            time_zone: tzid.clone(),
-        },
-    }
-}
+use crate::types::{Attendee, Event, EventStatus, EventTime, ParticipationStatus, Transparency};
 
 /// Convert our Event to a Google Calendar API Event
 pub fn to_google_event(event: &Event) -> google_calendar::types::Event {
@@ -46,11 +19,11 @@ pub fn to_google_event(event: &Event) -> google_calendar::types::Event {
     let reminders = if event.reminders.is_empty() {
         None
     } else {
-        Some(Reminders {
+        Some(google_calendar::types::Reminders {
             overrides: event
                 .reminders
                 .iter()
-                .map(|r| EventReminder {
+                .map(|r| google_calendar::types::EventReminder {
                     method: "popup".to_string(),
                     minutes: r.minutes,
                 })
@@ -59,26 +32,8 @@ pub fn to_google_event(event: &Event) -> google_calendar::types::Event {
         })
     };
 
-    let attendees: Vec<EventAttendee> = event
-        .attendees
-        .iter()
-        .map(|a| EventAttendee {
-            email: a.email.clone(),
-            display_name: a.name.clone().unwrap_or_default(),
-            response_status: a
-                .response_status
-                .map(participation_status_to_google)
-                .unwrap_or("needsAction")
-                .to_string(),
-            additional_guests: 0,
-            comment: String::new(),
-            id: String::new(),
-            optional: false,
-            organizer: false,
-            resource: false,
-            self_: false,
-        })
-        .collect();
+    let attendees: Vec<google_calendar::types::EventAttendee> =
+        event.attendees.iter().map(attendee_to_google).collect();
 
     let recurrence = event.recurrence.clone().unwrap_or_default();
 
@@ -99,6 +54,51 @@ pub fn to_google_event(event: &Event) -> google_calendar::types::Event {
         original_start_time,
         sequence: event.sequence.unwrap_or(0),
         ..Default::default()
+    }
+}
+
+fn attendee_to_google(attendee: &Attendee) -> google_calendar::types::EventAttendee {
+    google_calendar::types::EventAttendee {
+        email: attendee.email.clone(),
+        display_name: attendee.name.clone().unwrap_or_default(),
+        response_status: attendee
+            .response_status
+            .map(participation_status_to_google)
+            .unwrap_or("needsAction")
+            .to_string(),
+        additional_guests: 0,
+        comment: String::new(),
+        id: String::new(),
+        optional: false,
+        organizer: false,
+        resource: false,
+        self_: false,
+    }
+}
+
+/// Convert EventTime to Google's EventDateTime
+fn event_time_to_google(time: &EventTime) -> google_calendar::types::EventDateTime {
+    match time {
+        EventTime::Date(d) => google_calendar::types::EventDateTime {
+            date: Some(*d),
+            date_time: None,
+            time_zone: String::new(),
+        },
+        EventTime::DateTimeUtc(dt) => google_calendar::types::EventDateTime {
+            date: None,
+            date_time: Some(*dt),
+            time_zone: String::new(),
+        },
+        EventTime::DateTimeFloating(dt) => google_calendar::types::EventDateTime {
+            date: None,
+            date_time: Some(dt.and_utc()),
+            time_zone: String::new(),
+        },
+        EventTime::DateTimeZoned { datetime, tzid } => google_calendar::types::EventDateTime {
+            date: None,
+            date_time: Some(datetime.and_utc()),
+            time_zone: tzid.clone(),
+        },
     }
 }
